add query for all cases and case by ide
