+++ libs/api/domains/consultation-portal/src/lib/cases/cases.resolver.ts
@@ -9,6 +9,7 @@ import {
   FeatureFlag,
   Features,
 } from '@island.is/nest/feature-flags'
+import { GetCaseInput } from '../dto/case.input'
 
 @Resolver()
 @UseGuards(FeatureFlagGuard)
@@ -23,8 +24,10 @@ export class CaseResultResolver {
 
   @Query(() => CaseResult, { name: 'consultationPortalCaseById' })
   @FeatureFlag(Features.consultationPortalApplication)
-  async getCase(@Args('caseId') caseId: number): Promise<CaseResult> {
-    return await this.caseResultService.getCase(caseId)
+  async getCase(
+    @Args('input', { type: () => GetCaseInput }) input: GetCaseInput,
+  ): Promise<CaseResult> {
+    return await this.caseResultService.getCase(input)
   }
 
   @Query(() => [AdviceResult], { name: 'consultationPortalAdviceByCaseId' })
